Switch optimal_image to int and add instrumentation.

+++ optimal_image.h
@@ -3,5 +3,5 @@
 #include <time.h>
 #include "count.h"
 
-double *optimal_image(int width, int height, int wanted_width,
-	int wanted_height, double *image);
+int *optimal_image(int width, int height, int wanted_width,
+	int wanted_height, int *image);
